Refactor the test structure to avoid interferences among tests
