+++ tests/integration/istioctl/version_analyze/analyze_test.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package istioctl
+package version_analyze
 
 import (
 	"strings"
