+++ tests/integration/istioctl/version_analyze/testdata/reviews-svc.yaml
