+++ tests/integration/istioctl/version_analyze/testdata/reviews-vs-and-dr.yaml
