+++ tests/integration/istioctl/version_analyze/testdata/servicerole.yaml
