+++ tests/integration/istioctl/version_analyze/testdata/servicerolebinding.yaml
